tighten bytesource guarantees.
Byte sources are now required to return Exhausted as early as
possible. The guarantee we require from the byte source is that if at
any point of time a byte source returned `Remaining x`, then the
source should have at least one more byte left for the next call.  In
particular, if the source has `n` bytes left and we request for
exactly `n` bytes, then the result should be `Exhausted n`.

This is to ensure that primitives know when they are processing the
last block.