workaround a backpack bug 15138
In signature files, one cannot define a data with kindness nat.
(https://ghc.haskell.org/trac/ghc/ticket/15138). One of the
consequence of this workaround is that all implementations of
primitives in raaz should take 32-bit aligned pointers as
buffers. This is clearly undesirable and hence we should restore this
back to the original definition once the bug is fixed.