refactored hash memory type
Computing cryptographic hashes need to keep track of the

1. The hash of the data so far and,
2. The length of the data processed.

This commit introduces a new memory type that keeps track of these
data. There are two variants, one that keeps track of 64-bit length
quantities and the other that keeps track of 128 bit quantities. It is
a common memory type that can be used both for Blake2 and SHA2 hashes.