fix the buffer alignment field for the derived ones.
Ideally we should be able to just expose the BufferAlignment directly
from the base implementation for both the mac implementation and the
xchacha20. But for some reason I am unable to make it work with the
current version of backpack. While this needs to be investigated, we
commit a hack which will make the library work seamlessly for the
other implementations by hard-coding the alignment for these two
derived instances.